Avoid integer overflow warning in byte_cursor tests

+++ tests/cursor_test.c
@@ -87,7 +87,7 @@ static int s_test_byte_cursor_advance_internal(
     ASSERT_NOADVANCE(6, 5);
 
     ASSERT_NOADVANCE(SSIZE_MAX + 1, SSIZE_MAX);
-    ASSERT_NOADVANCE(SSIZE_MAX, SSIZE_MAX + 1);
+    ASSERT_NOADVANCE(SSIZE_MAX, (size_t)SSIZE_MAX + 1);
 
     return 0;
 }
